Con2Prim routine renaming
