+++ .github/workflows/github-actions-MacOS-clang.yml
@@ -67,7 +67,7 @@ jobs:
           md5sum -c hybrid_${{ matrix.test-function }}.md5
       - name: Run con2prim unit test for ${{ matrix.test-function }} function
         run: |
-          ./test/unit_test_Hybrid_${{ matrix.test-function }}
+          ./test/unit_test_hybrid_${{ matrix.test-function }}
 #      - name: Generate coverage report
 #        uses: ./.github/actions/code-coverage
