+++ .github/workflows/github-actions-MacOS-gcc.yml
@@ -65,7 +65,7 @@ jobs:
           curl -O https://raw.githubusercontent.com/SamuelCupp/GRHayL_TestData/main/con2prim/md5/hybrid_${{ matrix.test-function }}.md5
           md5sum -c hybrid_${{ matrix.test-function }}.md5
       - name: Run con2prim unit test for ${{ matrix.test-function }} function
-        run: ./test/unit_test_Hybrid_${{ matrix.test-function }}
+        run: ./test/unit_test_hybrid_${{ matrix.test-function }}
 #      - name: Generate coverage report
 #        uses: ./.github/actions/code-coverage
