+++ GRHayL/Con2Prim/Hybrid/FontFix/make.code.defn
@@ -1,5 +1,5 @@
 # Subdirectory make.code.defn for Con2Prim/Font_Fix
 
 # Source files in this directory that need compiled:
-SRCS = Hybrid_Font_Fix.c \
-       hybrid_font_fix_loop.c
+SRCS = hybrid_Font_fix.c \
+       hybrid_Font_fix_loop.c
