+++ GRHayL/Con2Prim/Hybrid/GNR1D.c
@@ -1,6 +1,6 @@
 #include "../harm_u2p_util.h"
 
-int newton_raphson_1d(
+int grhayl_newton_raphson_1d(
       const eos_parameters *restrict eos,
       const harm_aux_vars_struct *restrict harm_aux,
       const int ndim,
@@ -15,8 +15,6 @@ int newton_raphson_1d(
 
   // Initialize various parameters and variables:
   int n_iter = 0;
-  int i_extra = 0;
-  int doing_extra = 0;
   double errx = 1.0;
   double df = 1.0;
   double f = 1.0;
@@ -34,25 +32,14 @@ int newton_raphson_1d(
     /****************************************/
     /* Calculate the convergence criterion */
     /****************************************/
-    errx  = (x[0]==0.0) ? fabs(dx[0]) : fabs(dx[0]/x[0]);
+    errx = (x[0]==0.0) ? fabs(dx[0]) : fabs(dx[0]/x[0]);
 
     /****************************************/
     /* Make sure that the new x[] is physical : */
     /****************************************/
     x[0] = fabs(x[0]);
 
-    /*****************************************************************************/
-    /* If we've reached the tolerance level, then just do a few extra iterations */
-    /*  before stopping                                                          */
-    /*****************************************************************************/
-    if( (fabs(errx) <= NEWT_TOL) && (doing_extra == 0) && (EXTRA_NEWT_ITER > 0) ) {
-      doing_extra = 1;
-    }
-
-    if( doing_extra == 1 ) i_extra++;
-
-    if( ((fabs(errx) <= NEWT_TOL)&&(doing_extra == 0))
-        || (i_extra > EXTRA_NEWT_ITER) || (n_iter >= (MAX_NEWT_ITER-1)) ) {
+    if( ((fabs(errx) <= NEWT_TOL)) || (n_iter >= (MAX_NEWT_ITER-1)) ) {
       keep_iterating = 0;
     }
 
@@ -63,19 +50,11 @@ int newton_raphson_1d(
   *n_iter_ptr = n_iter;
 
   /*  Check for bad untrapped divergences : */
-//TODO: error-checking
-//  if( (robust_isfinite(f)==0) ||  (robust_isfinite(df)==0) ) {
-//    return(2);
-//  }
-
-  if( fabs(errx) <= NEWT_TOL ) {
-    return(0);
-  } else if( (fabs(errx) <= MIN_NEWT_TOL) && (fabs(errx) > NEWT_TOL) ) {
-    return(0);
+  if( !isfinite(f) ||  !isfinite(df) ) {
+    return 3;
+  } else if( fabs(errx) <= NEWT_TOL ) {
+    return 0;
   } else {
-    return(1);
+    return 2;
   }
-
-  return(0);
-
 }
