+++ GRHayL/Con2Prim/Hybrid/Noble2D/GNR2D.c
@@ -1,8 +1,6 @@
 #include "../../harm_u2p_util.h"
 
-void validate_x(double x[2], const double x0[2] );
-
-int general_newton_raphson(
+int grhayl_general_newton_raphson(
       const eos_parameters *restrict eos,
       const harm_aux_vars_struct *restrict harm_aux,
       const int ndim,
@@ -25,7 +23,7 @@ int general_newton_raphson(
   /* Start the Newton-Raphson iterations : */
   int keep_iterating = 1;
   while( keep_iterating ) {
-    funcd(eos, harm_aux, ndim, indep_var_in, x, dx, resid, jac, &f, &df, n_iter_ptr);  /* returns with new dx, f, df */
+    funcd(eos, harm_aux, ndim, indep_var_in, x, dx, resid, jac, &f, &df, n_iter_ptr);
 
     /* Save old values before calculating the new: */
     errx = 0.0;
@@ -46,7 +44,7 @@ int general_newton_raphson(
     /****************************************/
     /* Make sure that the new x[] is physical : */
     /****************************************/
-    validate_x(x, x_old);
+    grhayl_validate_x(x, x_old);
 
     if( ((fabs(errx) <= NEWT_TOL)) || (n_iter >= (MAX_NEWT_ITER-1)) ) {
       keep_iterating = 0;
@@ -66,15 +64,4 @@ int general_newton_raphson(
   } else {
     return 2;
   }
-}
-
-void validate_x(double x[2], const double x0[2]) {
-  const double dv = 1.e-15;
-
-  /* Always take the absolute value of x[0] and check to see if it's too big:  */
-  x[0] = fabs(x[0]);
-  x[0] = (x[0] > Z_TOO_BIG) ?  x0[0] : x[0];
-
-  x[1] = (x[1] < 0.) ?   0.       : x[1];  /* if it's too small */
-  x[1] = (x[1] > 1.) ?  (1. - dv) : x[1];  /* if it's too big   */
 }
