+++ GRHayL/Con2Prim/Hybrid/Noble2D/make.code.defn
@@ -1,6 +1,6 @@
 # Subdirectory make.code.defn for Con2Prim/C2P_Routines/Hybrid/Noble2D
 
 # Source files in this directory that need compiled:
-SRCS = func_vsq.c       \
-       GNR2D.c          \
-       Hybrid_Noble2D.c
+SRCS = func_vsq.c \
+       GNR2D.c    \
+       hybrid_Noble2D.c
