+++ GRHayL/Con2Prim/Hybrid/func_rho.c
@@ -23,7 +23,7 @@
           D   = W * rho, set in the Utoprim_new_body() functoin
 *********************************************************************************/
 // for the isentropic version: eq. (27)
-void func_rho(
+void grhayl_func_rho(
       const eos_parameters *restrict eos,
       const harm_aux_vars_struct *restrict harm_aux,
       const int ndim,
@@ -61,6 +61,4 @@ void func_rho(
   dx[0] = - resid[0]/jac[0][0];
   *df   = - resid[0]*resid[0];
   *f    = - 0.5*(*df);
-
-  return;
 }
