+++ GRHayL/Con2Prim/Tabulated/Palenzuela1D/make.code.defn
@@ -1,3 +1,3 @@
-SRCS = Tabulated_Palenzuela1D.c        \
-       Tabulated_Palenzuela1D_energy.c \
-       Tabulated_Palenzuela1D_entropy.c
+SRCS = tabulated_Palenzuela1D.c        \
+       tabulated_Palenzuela1D_energy.c \
+       tabulated_Palenzuela1D_entropy.c
