+++ GRHayL/Con2Prim/Tabulated/Palenzuela1D/tabulated_Palenzuela1D_entropy.c
@@ -1,12 +1,11 @@
 #include "../utils.h"
 
 /*
- * Function   : compute_rho_P_eps_T_W_energy
+ * Function   : compute_rho_P_eps_T_W_entropy
  * Author     : Leo Werneck
  *
  * This function is used to compute rho, P, eps, T, and W from x and the input
- * conservative variables, using the specific internal energy to recover the
- * temperature.
+ * conservative variables, using the specific entropy to recover the temperature.
  *
  * Parameters : x       - Current guess for the root of f(x).
  *            : fparams - Parameters used to evaluate f(x) (see palenzuela.h).
@@ -20,9 +19,10 @@
  *
  * References : [1] Palenzuela et al. (2015) arXiv:1505.01607
  *            : [2] Siegel et al. (2018) arXiv:1712.07538
+ *            : [3] Werneck et al. (2023) arXiv:2208.14487
  */
 static void
-compute_rho_P_eps_T_W_energy(
+compute_rho_P_eps_T_W_entropy(
       const double x,
       fparams_struct *restrict fparams,
       double *restrict rho_ptr,
@@ -43,12 +43,8 @@ compute_rho_P_eps_T_W_energy(
   // Step 2: Now compute P and eps
   double P, eps;
   if( fparams->evolve_T ) {
-    const double q = fparams->q;
-    const double s = fparams->s;
-    const double t = fparams->t;
-    // Eq. (43) of https://arxiv.org/pdf/1712.07538.pdf
-    eps = W - 1.0 + (1.0-W*W)*x/W + W*(q - s + t*t/(2*x*x) + s/(2*W*W) );
-    eos->tabulated_compute_P_T_from_eps( eos, rho, Y_e, eps, &P, &T );
+    double ent = fparams->cons_undens->entropy/W;
+    eos->tabulated_compute_P_eps_T_from_S( eos, rho, Y_e, ent, &P, &eps, &T );
   }
   else {
     // If the temperature is not evolved, use the input guess to determine
@@ -66,15 +62,15 @@ compute_rho_P_eps_T_W_energy(
 }
 
 /*
- * Function : Tabulated_Palenzuela1D_energy
+ * Function : grhayl_tabulated_Palenzuela1D_entropy
  * Author   : Leo Werneck
  *
- * This is a wrapper function around the Tabulated_Palenzuela1D that performs
+ * This is a wrapper function around the grhayl_tabulated_Palenzuela1D that performs
  * a primitive recovery using the Palenzuela et al. scheme using the specific
- * energy to recover the temperature. See file Tabulated_Palenzuela1D.c for
+ * entropy to recover the temperature. See file grhayl_tabulated_Palenzuela1D.c for
  * further details.
  */
-int Tabulated_Palenzuela1D_energy(
+int grhayl_tabulated_Palenzuela1D_entropy(
       const GRHayL_parameters *restrict grhayl_params,
       const eos_parameters *restrict eos,
       const metric_quantities *restrict ADM_metric,
@@ -83,11 +79,11 @@ int Tabulated_Palenzuela1D_energy(
       primitive_quantities *restrict prims,
       con2prim_diagnostics *restrict diagnostics ) {
 
-  return Tabulated_Palenzuela1D(compute_rho_P_eps_T_W_energy,
-                                grhayl_params,
-                                eos,
-                                ADM_metric,
-                                cons_undens,
-                                prims,
-                                diagnostics);
+  return grhayl_tabulated_Palenzuela1D( compute_rho_P_eps_T_W_entropy,
+                                 grhayl_params,
+                                 eos,
+                                 ADM_metric,
+                                 cons_undens,
+                                 prims,
+                                 diagnostics );
 }
