+++ GRHayL/Con2Prim/Tabulated/toms748.c
@@ -210,7 +210,7 @@ cubic_interpolate(
 }
 
 roots_error_t
-toms748(
+grhayl_toms748(
     double f(const double, void *restrict),
     void *restrict fparams,
     double a,
