+++ implementations/GRHayLHD/src/convert_HydroBase_to_GRHayLHD.c
@@ -29,14 +29,6 @@ void convert_HydroBase_to_GRHayLHD(CCTK_ARGUMENTS) {
       for(int i=0; i<imax; i++) {
         const int index=CCTK_GFINDEX3D(cctkGH,i,j,k);
 
-        //TODO: this references simple gamma law; clearly needs to be extended to more EOS types
-        // P = (\Gamma - 1) rho epsilon
-        // -> \Gamma = P/(rho epsilon) + 1
-        const double measured_gamma = ( press[index]/(rho[index] * eps[index]) + 1.0 );
-        if(rho[index]>grhayl_eos->rho_atm && fabs(grhayl_eos->Gamma_th - measured_gamma)/grhayl_eos->Gamma_th > 1e-2)
-          CCTK_VERROR("Expected simple gamma law with gamma_th=%.15e, but found a point with gamma law such that gamma_th=%.15e. error = %e| rb=%e rbatm=%e P=%e\n",
-                      grhayl_eos->Gamma_th, measured_gamma, (grhayl_eos->Gamma_th-measured_gamma)/grhayl_eos->Gamma_th, rho[index], grhayl_eos->rho_atm, press[index] );
-
         rho_b[index] = rho[index];
         pressure[index] = press[index];
